pyproject.toml
[project]
name = "embedding-visualizer"
version = "0.1.0"
description = "Interactive visualization of text embeddings with dimensionality reduction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clerk-backend-api>=1.8.0",
    "fastapi[standard]>=0.115.11",
    "pydantic>=2.10.6",
    "sentence-transformers>=3.4.1",
    "scikit-learn>=1.6.1",
    "numpy>=1.23.5,<1.24.0",
    "umap-learn>=0.5.7",
    "openTSNE>=1.0.2",
    "posthog>=3.18.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "redis>=5.2.1",
    "pydantic-settings>=2.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "ruff>=0.9.9",
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
]

[tool.mypy]
ignore_missing_imports = true
